requirements.txt
blinker==1.9.0
cairocffi==1.7.1
CairoSVG==2.7.1
cffi==1.17.1
chardet==5.2.0
click==8.1.8
colorama==0.4.6
cssselect2==0.7.0
defusedxml==0.7.1
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
magickwand==0.2
MarkupSafe==2.1.5
pdf2image==1.17.0
pillow==11.1.0
pycparser==2.22
python-barcode==0.15.1
qrcode==8.0
reportlab==4.2.5
SQLAlchemy==2.0.36
tinycss2==1.4.0
treepoem==3.25.0
typing_extensions==4.12.2
Wand==0.6.13
webencodings==0.5.1
Werkzeug==3.1.3
WTForms==3.2.1
